+++ R/basilisk.R
@@ -1,5 +1,5 @@
 anndata_env <- basilisk::BasiliskEnvironment(
     envname="anndata_env",
     pkgname="zellkonverter",
-    packages=c("pandas==0.25.1", "sklearn==0.22.1")
+    packages=c("anndata==0.7.3")
 )